+++ CHANGELOG.md
@@ -4,7 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [[0.1.0] 2017-09-25](https://github.com/arturock/vscode-gitstash/tree/v0.2.0)
+## [[0.3.0] 2017-09-26](https://github.com/arturock/vscode-gitstash/tree/v0.3.0)
+- Configure diff view title
+- Improve stash watcher
+
+## [[0.2.0] 2017-09-25](https://github.com/arturock/vscode-gitstash/tree/v0.2.0)
 - Add configurations for tree items
 
 ## [[0.1.0] 2017-09-24](https://github.com/arturock/vscode-gitstash/tree/v0.1.0)